refactor: move authContext to auth folder
