+++ src/templates/Header/mobile.tsx
@@ -6,7 +6,7 @@ import Navigation from '../../components/Navigation';
 import MenuIcon from '@material-ui/icons/Menu';
 import PersonAddIcon from '@material-ui/icons/PersonAdd';
 import { useState } from 'react';
-import { useAuthContextState } from '../../context/authContext';
+import { useAuthContextState } from '../../context/auth/authContext';
 import { NavigationProps } from './types';
 
 const useStyles = makeStyles((theme: Theme) =>
