+++ src/templates/Main/index.tsx
@@ -2,8 +2,8 @@ import { ApolloClient, createHttpLink, InMemoryCache, ApolloProvider } from '@ap
 import { ThemeProvider } from '@material-ui/styles';
 import MomentUtils from '@date-io/moment';
 import 'moment/locale/pl';
-import { AuthProvider } from '../../context/authContext';
-import LocationProvider from '../../context/locationContext/locationContext';
+import { AuthProvider } from '../../context/auth/authContext';
+import LocationProvider from '../../context/location/locationContext';
 import GlobalStyles from '../../theme/globalStyles';
 import theme from '../../theme/material-theme';
 import { SnackbarProvider } from 'notistack';
